Emerald: Speaker uses Speech Service
This change fixes a problem that was there since the very beginning
of the Android version: interpreting a Rexx was taking 3secs at the start.

Thanks to the Profiler for having detected a doing-nothing in the middle
of nowhere, and thanks to the Debugger for having found that 'nowhere'
was for closing the "Speaker".

First fix consisted to flush the Rexx buffer before to close the Speaker.
But that wasn't enough.  Mimicing what I had done for (Crossper) communication
service, I rewrote the Speaker as a client of a Speech Service.

Now, YAXX is really responsive, as it was used to be.

Also, this change announces the 'Emerald Peak' revision. It was initially
started to ease Yaxx at dealing with the Android system -- shelling a REXX
script from REXX.  This isn't yet done, but will soon.