Init and finalize Rexx context, n times interpreting in the meanwhile.
Following the "activity-single-top" change, `init' and `finalize' were required
to be able to shell Rexx from an already alive Rexx, while keeping
same SystemContext: JConsole, Speaker, files, etc.

Each "interpret" runs on its own thread, and I realized that the JNIEnv pointer
was being modified while switching threads, as well as the JVM reorganizing the
addresses of the objects (after a Garbage Collection.)

For the former, I created a global field g_jniEnv, set at each `interpret' call,
reset at the end.  I'm not relying anymore in a same env pointer being part of
any instances that need it: they will use the volatile g_jniEnv field.

For the latter, I had to pin down the instances (Speaker, JConsole) to make
the JVM consider them as "unmoveable" addresses, that is using NewGlobalRef
the address of which is deemed to stay constant among threads.