Not too sure on how useful those changes are...
I'm still chasing the erratic SIGSEGV sometimes occurring when pressing
the back button. I was thinking the Speech class was the culprit; I revamped
the Speech termination, and, IMHO these changes are rather better, or
at least harmless. But this is not it.

It looks like ScriptsList that starts Rexx may be destroyed
before Rexx is destroyed. This dangling order might be at the
origin of the problem, and I will try to fix it in next commit.