Fix for problems in REXX Activity lifecycle.
Finally!  This code seems to have robust manners to keep a sane SystemContext
among the several ways Rexx can be called. The solution was to use the
couple onStart/onStop rather than onCreate/onDestroy.

It might look obvious, but it wasn't. I spent 2 days before to arrive to
this solution.  I tried a Ref'ed SystemContext, and to make Rexx single task
(but this doesn't work, because Google decided to secure access to GMail
attachments, for onNewIntend.)  I even wrote a LocalBroadcaster! Nothing
worked, a pain in the neck at fighting SIGSEGV's all over the place.

Rereading carefully the Activities'lifecycle -- despite Google efforts for
documenting, this part is still unclear for many -- I found out that the
onStart/onStop had to recreate the SystemContext, b/c I could not keep
the same context for Rexx started from the App, and Rexx started as an intent
for executing mail attachment, or file association.

REXX'SystemContext is really what it says it is: independent of REXX'guts
(that is: clauses, the stack, etc..), it's just a kind of Helper to redirect
the output/input (serambuffers) resources to what's appropriate in a given
(new) context.