Branch b2
Getting rid of Launcher + ScriptsList activities.
The Launcher was there only for switching views (splash -> list).
Use a single activity (Landing) instead, with a ViewAnimator.

Then..  A severe problem was discovered.  It stops the Rexx Interpreter
(Tokenizer?) at the very start. Current change doesn't appear to be the
culprit, and I ever rolled back to July 22nd and the problem is still there.

This means that it should be something in the configuration (compiler, build
tools.)  In any case, I cannot pursue on 'emerald' anymore, and created the
branch 'b2'.