And the culprit was...
Proguard! It's a bit my fault: I forgot to tell that "context", a long
value in Rexx.java, should not be obfuscated.  But, even after
this change + new proguard CFG structure, the problem reappeared.

No proguard, no problemo.  Proguard is less than 1% of footprint improvement.
Why would I to bother for this ridiculous amount? I'm not even trying to
spend one second at fixing my Proguard config.  That was enough a pain
in the neck, and I just don't care.

This is branch b2, soon to be merged back to Emerald.